Isolate macro types
